Add Faculty class with properties and licensing info
This commit introduces a new `Faculty` class in the
`UniversitySystem.Domain.Entities` namespace. The class
inherits from `BaseEntity` and includes properties for
`Name`, `UniversityId`, a navigation property for
`University`, and a collection of `Departments`. A
copyright notice and MIT License information have also
been added at the top of the file.